+++ vlingo-reactive-messaging-patterns/src/main/java/io/vlingo/reactive/messaging/patterns/scattergather/BestPriceQuotation.java
@@ -43,59 +43,4 @@ public class BestPriceQuotation
         }
         return priceQuotes.size() > 0 ? String.format( sb.toString(), quotes ) : sb.toString();
     }
-    
-    public static class PriceQuote
-    {
-        public final QuoteProcessor quoteProcessor;
-        public final String rfqId;
-        public final String itemId;
-        public final Double retailPrice;
-        public final Double discountPrice;
-        
-        public PriceQuote( final QuoteProcessor quoteProcessor, final String rfqId, final String itemId, final Double retailPrice, final Double discountPrice )
-        {
-            this.quoteProcessor = quoteProcessor;
-            this.rfqId = rfqId;
-            this.itemId = itemId;
-            this.retailPrice = retailPrice;
-            this.discountPrice = discountPrice;
-        }
-
-        /* @see java.lang.Object#toString() */
-        @Override
-        public String toString()
-        {
-            return String.format( "%s %s %s $%.2f $%.2f", quoteProcessor.getClass().getSimpleName(), rfqId, itemId, retailPrice, discountPrice );
-        }
-    }
-    
-    public static class QuotationFulfillment
-    {
-        public final String rfqId;
-        public final Integer quotesRequested;
-        public final Vector<PriceQuote> priceQuotes;
-        public final OrderProcessor requestor;
-        
-        public QuotationFulfillment( final String rfqId, Integer quotesRequested, OrderProcessor requestor, PriceQuote...quotes )
-        {
-            this.rfqId = rfqId;
-            this.quotesRequested = quotesRequested;
-            this.requestor = requestor;
-            this.priceQuotes = new Vector<>();
-            for ( PriceQuote quote : quotes )
-            {
-                priceQuotes.add( quote );
-            }
-        }
-    }
-    
-    public static class PriceQuoteTimedOut
-    {
-        public final String rfqId;
-        
-        public PriceQuoteTimedOut( String rfqId )
-        {
-            this.rfqId = rfqId;
-        }
-    }
 }
