+++ vlingo-reactive-messaging-patterns/src/main/java/io/vlingo/reactive/messaging/patterns/scattergather/PriceQuoteTimedOut.java
@@ -1,0 +1,20 @@
+// Copyright © 2012-2018 Vaughn Vernon. All rights reserved.
+//
+// This Source Code Form is subject to the terms of the
+// Mozilla Public License, v. 2.0. If a copy of the MPL
+// was not distributed with this file, You can obtain
+// one at https://mozilla.org/MPL/2.0/.
+package io.vlingo.reactive.messaging.patterns.scattergather;
+
+/**
+ * PriceQuoteTimedOut
+ */
+public class PriceQuoteTimedOut
+{
+    public final String rfqId;
+    
+    public PriceQuoteTimedOut( String rfqId )
+    {
+        this.rfqId = rfqId;
+    }
+}
