+++ vlingo-reactive-messaging-patterns/src/main/java/io/vlingo/reactive/messaging/patterns/scattergather/RetailItem.java
@@ -1,0 +1,22 @@
+// Copyright © 2012-2018 Vaughn Vernon. All rights reserved.
+//
+// This Source Code Form is subject to the terms of the
+// Mozilla Public License, v. 2.0. If a copy of the MPL
+// was not distributed with this file, You can obtain
+// one at https://mozilla.org/MPL/2.0/.
+package io.vlingo.reactive.messaging.patterns.scattergather;
+
+/**
+ * RetailItem
+ */
+public class RetailItem
+{
+    public final String itemId;
+    public final Double retailPrice;
+    
+    public RetailItem( final String itemId, final Double retailPrice )
+    {
+        this.itemId = itemId;
+        this.retailPrice = retailPrice;
+    }
+}
